requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   openai
asgiref==3.11.0
    # via django
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
colorama==0.4.6
    # via tqdm
distro==1.9.0
    # via openai
django==6.0.1
    # via
    #   -r requirements.in
    #   django-extensions
    #   djangorestframework
django-extensions==4.1
    # via -r requirements.in
djangorestframework==3.16.1
    # via -r requirements.in
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   openai
    #   python-telegram-bot
idna==3.11
    # via
    #   anyio
    #   httpx
jiter==0.13.0
    # via openai
openai==2.16.0
    # via -r requirements.in
pydantic==2.12.5
    # via openai
pydantic-core==2.41.5
    # via pydantic
python-decouple==3.8
    # via -r requirements.in
python-telegram-bot==22.6
    # via -r requirements.in
sniffio==1.3.1
    # via openai
sqlparse==0.5.5
    # via django
tqdm==4.67.2
    # via openai
typing-extensions==4.15.0
    # via
    #   openai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via django

requests